requirements.txt
# --- Core runtime ---
requests>=2.32,<3
urllib3>=2.0,<3
python-dotenv>=1.2,<2

# --- Data & numerics ---
numpy>=2.0,<3
pandas>=2.3,<3
scipy>=1.16,<2

# --- Validation & typing ---
jsonschema>=4.25,<5
typing_extensions>=4.14,<5

# --- Utilities ---
tqdm>=4.67,<5
